refactor(git): split status.ts into source-control modules
