+++ config/max-lines-baseline.txt
@@ -20,7 +20,6 @@ inline src/main/codex-accounts/service.ts
 inline src/main/codex/hook-service.ts
 inline src/main/daemon/daemon-init.ts
 inline src/main/git/runner.ts
-inline src/main/git/status.ts
 inline src/main/git/worktree.ts
 inline src/main/github/project-view.ts
 inline src/main/index.ts
