+++ src/main/git/source-control/branch-change-entries.ts
@@ -1,0 +1,145 @@
+import type { GitBranchChangeEntry } from '../../../shared/git-diff-compare-types'
+import type { GitBranchChangeStatus } from '../../../shared/git-status-types'
+import { parseNumstat } from '../../../shared/git-uncommitted-line-stats'
+import { decodeGitCQuotedPath } from '../../../shared/git-cquoted-path'
+import type { GitRuntimeOptions } from '../git-runtime-options'
+import { gitOptionsForWorktree } from '../git-runtime-options'
+import { gitExecFileAsync } from '../runner'
+import { MAX_GIT_SHOW_BYTES } from './git-show-max-bytes'
+
+function parseBranchStatusChar(char: string): GitBranchChangeStatus {
+  switch (char) {
+    case 'M':
+      return 'modified'
+    case 'A':
+      return 'added'
+    case 'D':
+      return 'deleted'
+    case 'R':
+      return 'renamed'
+    case 'C':
+      return 'copied'
+    default:
+      return 'modified'
+  }
+}
+
+export async function loadBranchChanges(
+  worktreePath: string,
+  mergeBase: string,
+  headOid: string,
+  options: GitRuntimeOptions = {}
+): Promise<GitBranchChangeEntry[]> {
+  // Why: core.quotePath=false keeps real UTF-8 paths — see getStatus rationale.
+  const gitOptions = {
+    ...gitOptionsForWorktree(worktreePath, options),
+    maxBuffer: MAX_GIT_SHOW_BYTES
+  }
+  // Why: both diffs are independent, so run them concurrently instead of serializing.
+  const [{ stdout }, { stdout: numstat }] = await Promise.all([
+    gitExecFileAsync(
+      ['-c', 'core.quotePath=false', 'diff', '--name-status', '-M', '-C', mergeBase, headOid],
+      gitOptions
+    ),
+    gitExecFileAsync(
+      ['-c', 'core.quotePath=false', 'diff', '-z', '--numstat', '-M', '-C', mergeBase, headOid],
+      gitOptions
+    )
+  ])
+  const statsByPath = parseNumstat(numstat)
+
+  const entries: GitBranchChangeEntry[] = []
+  // Why: split on /\r?\n/ so Git's CRLF output on Windows leaves no trailing \r in paths.
+  for (const line of stdout.split(/\r?\n/)) {
+    if (!line) {
+      continue
+    }
+    const entry = parseBranchChangeLine(line)
+    if (entry) {
+      entries.push({ ...entry, ...statsByPath.get(entry.path) })
+    }
+  }
+  return entries
+}
+
+export async function loadCommitChanges(
+  worktreePath: string,
+  parentOid: string | null,
+  commitOid: string,
+  options: GitRuntimeOptions = {}
+): Promise<GitBranchChangeEntry[]> {
+  // Why: root commits have no parent tree; diff-tree --root uses git's empty tree, avoiding a hardcoded hash-format-specific oid.
+  const args = parentOid
+    ? ['-c', 'core.quotePath=false', 'diff', '--name-status', '-M', '-C', parentOid, commitOid]
+    : [
+        '-c',
+        'core.quotePath=false',
+        'diff-tree',
+        '--root',
+        '--no-commit-id',
+        '--name-status',
+        '-r',
+        '-M',
+        '-C',
+        commitOid
+      ]
+  const numstatArgs = parentOid
+    ? ['-c', 'core.quotePath=false', 'diff', '-z', '--numstat', '-M', '-C', parentOid, commitOid]
+    : [
+        '-c',
+        'core.quotePath=false',
+        'diff-tree',
+        '-z',
+        '--root',
+        '--no-commit-id',
+        '--numstat',
+        '-r',
+        '-M',
+        '-C',
+        commitOid
+      ]
+  const gitOptions = {
+    ...gitOptionsForWorktree(worktreePath, options),
+    maxBuffer: MAX_GIT_SHOW_BYTES
+  }
+  // Why: the two git queries are independent, so run them in parallel.
+  const [{ stdout }, { stdout: numstat }] = await Promise.all([
+    gitExecFileAsync(args, gitOptions),
+    gitExecFileAsync(numstatArgs, gitOptions)
+  ])
+  const statsByPath = parseNumstat(numstat)
+
+  const entries: GitBranchChangeEntry[] = []
+  for (const line of stdout.split(/\r?\n/)) {
+    if (!line) {
+      continue
+    }
+    const entry = parseBranchChangeLine(line)
+    if (entry) {
+      entries.push({ ...entry, ...statsByPath.get(entry.path) })
+    }
+  }
+  return entries
+}
+
+export function parseBranchChangeLine(line: string): GitBranchChangeEntry | null {
+  const parts = line.split('\t')
+  const rawStatus = parts[0] ?? ''
+  const status = parseBranchStatusChar(rawStatus[0] ?? 'M')
+
+  if (rawStatus.startsWith('R') || rawStatus.startsWith('C')) {
+    const oldPath = decodeGitCQuotedPath(parts[1] ?? '')
+    const path = decodeGitCQuotedPath(parts[2] ?? '')
+    if (!path) {
+      return null
+    }
+    return { path, oldPath, status }
+  }
+
+  const path = decodeGitCQuotedPath(parts[1] ?? '')
+  if (!path) {
+    return null
+  }
+
+  return { path, status }
+}
