+++ src/main/git/source-control/commit-diff.ts
@@ -1,0 +1,70 @@
+import type { GitDiffResult } from '../../../shared/git-diff-compare-types'
+import { stableInFlightKey } from '../../../shared/in-flight-promise-dedupe'
+import type { GitRuntimeOptions } from '../git-runtime-options'
+import { gitRuntimeOptionsKey } from './git-runtime-options-cache-key'
+import { gitDiffReadDedupe } from './git-read-cache-invalidation'
+import { buildDiffResult } from './diff-result'
+import { readGitBlobAtOidPath } from './git-blob-read'
+
+export async function getCommitDiff(
+  worktreePath: string,
+  args: {
+    commitOid: string
+    parentOid?: string | null
+    filePath: string
+    oldPath?: string
+  },
+  options: GitRuntimeOptions = {}
+): Promise<GitDiffResult> {
+  return gitDiffReadDedupe.run(
+    stableInFlightKey([
+      'commitDiff',
+      worktreePath,
+      args.commitOid,
+      args.parentOid ?? null,
+      args.filePath,
+      args.oldPath ?? null,
+      ...gitRuntimeOptionsKey(options)
+    ]),
+    () => loadCommitDiff(worktreePath, args, options)
+  )
+}
+
+async function loadCommitDiff(
+  worktreePath: string,
+  args: {
+    commitOid: string
+    parentOid?: string | null
+    filePath: string
+    oldPath?: string
+  },
+  options: GitRuntimeOptions
+): Promise<GitDiffResult> {
+  try {
+    const leftPath = args.oldPath ?? args.filePath
+    // Why concurrent: the two sides are independent `git show` spawns. A root
+    // commit has no parent to read, so that side resolves without a spawn.
+    const [leftBlob, rightBlob] = await Promise.all([
+      args.parentOid
+        ? readGitBlobAtOidPath(worktreePath, args.parentOid, leftPath, options)
+        : Promise.resolve({ content: '', isBinary: false }),
+      readGitBlobAtOidPath(worktreePath, args.commitOid, args.filePath, options)
+    ])
+
+    return buildDiffResult(
+      leftBlob.content,
+      rightBlob.content,
+      leftBlob.isBinary,
+      rightBlob.isBinary,
+      args.filePath
+    )
+  } catch {
+    return {
+      kind: 'text',
+      originalContent: '',
+      modifiedContent: '',
+      originalIsBinary: false,
+      modifiedIsBinary: false
+    }
+  }
+}
