+++ src/main/git/source-control/git-blob-read.ts
@@ -1,0 +1,121 @@
+import { readFile, stat } from 'node:fs/promises'
+import * as path from 'node:path'
+import { isBinaryBuffer } from '../../../shared/binary-buffer'
+import type { GitRuntimeOptions } from '../git-runtime-options'
+import { gitOptionsForWorktree } from '../git-runtime-options'
+import { gitExecFileAsyncBuffer } from '../runner'
+import { isMaxBufferOverflowError } from '../max-buffer-overflow'
+import { MAX_GIT_SHOW_BYTES } from './git-show-max-bytes'
+import { PREVIEWABLE_BINARY_MIME_TYPES } from './previewable-binary-mime-types'
+
+export type GitBlobReadResult = {
+  content: string
+  isBinary: boolean
+  exists: boolean
+}
+
+export async function readUnstagedLeftBlob(
+  worktreePath: string,
+  filePath: string,
+  options: GitRuntimeOptions = {}
+): Promise<GitBlobReadResult> {
+  const indexBlob = await readGitBlobAtIndexPath(worktreePath, filePath, options)
+  if (indexBlob.exists) {
+    return indexBlob
+  }
+
+  return readGitBlobAtOidPath(worktreePath, 'HEAD', filePath, options)
+}
+
+export async function readGitBlobAtIndexPath(
+  worktreePath: string,
+  filePath: string,
+  options: GitRuntimeOptions = {}
+): Promise<GitBlobReadResult> {
+  // Why: Git's `:<path>` syntax expects forward slashes even on Windows.
+  const gitPath = filePath.replace(/\\/g, '/')
+  try {
+    const { stdout } = await gitExecFileAsyncBuffer(['show', `:${gitPath}`], {
+      ...gitOptionsForWorktree(worktreePath, options),
+      maxBuffer: MAX_GIT_SHOW_BYTES
+    })
+
+    return { ...bufferToBlob(stdout, filePath), exists: true }
+  } catch (error) {
+    if (isMaxBufferOverflowError(error)) {
+      return { content: '', isBinary: true, exists: true }
+    }
+    return { content: '', isBinary: false, exists: false }
+  }
+}
+
+export async function readGitBlobAtOidPath(
+  worktreePath: string,
+  oid: string,
+  filePath: string,
+  options: GitRuntimeOptions = {}
+): Promise<GitBlobReadResult> {
+  // Why: Git's `<oid>:<path>` syntax expects forward slashes even on Windows.
+  const gitPath = filePath.replace(/\\/g, '/')
+  try {
+    const { stdout } = await gitExecFileAsyncBuffer(
+      ['show', '--end-of-options', `${oid}:${gitPath}`],
+      {
+        ...gitOptionsForWorktree(worktreePath, options),
+        maxBuffer: MAX_GIT_SHOW_BYTES
+      }
+    )
+
+    return { ...bufferToBlob(stdout, filePath), exists: true }
+  } catch (error) {
+    if (isMaxBufferOverflowError(error)) {
+      return { content: '', isBinary: true, exists: true }
+    }
+    return { content: '', isBinary: false, exists: false }
+  }
+}
+
+export async function readWorkingTreeFile(filePath: string): Promise<GitBlobReadResult> {
+  let fileStat
+  try {
+    fileStat = await stat(filePath)
+  } catch (error) {
+    // Why: only ENOENT is a real deletion; other stat errors are read failures, not absence.
+    return {
+      content: '',
+      isBinary: false,
+      exists: (error as NodeJS.ErrnoException)?.code !== 'ENOENT'
+    }
+  }
+  if (!fileStat.isFile()) {
+    return { content: '', isBinary: false, exists: false }
+  }
+  if (fileStat.size > MAX_GIT_SHOW_BYTES) {
+    // Why: mirror git's maxBuffer cap for working-tree reads so readFile can't pull in huge assets.
+    return { content: '', isBinary: true, exists: true }
+  }
+  try {
+    const buffer = await readFile(filePath)
+    return bufferToBlob(buffer, filePath)
+  } catch {
+    // Why: the file exists but could not be read — a read failure, not a deletion.
+    return { content: '', isBinary: false, exists: true }
+  }
+}
+
+function bufferToBlob(buffer: Buffer, filePath?: string): GitBlobReadResult {
+  const isBinary = isBinaryBuffer(buffer)
+  // Return base64 for recognized image formats so the renderer can display them
+  const isPreviewableBinary = filePath
+    ? !!PREVIEWABLE_BINARY_MIME_TYPES[path.extname(filePath).toLowerCase()]
+    : false
+  return {
+    content: isBinary
+      ? isPreviewableBinary
+        ? buffer.toString('base64')
+        : ''
+      : buffer.toString('utf-8'),
+    isBinary,
+    exists: true
+  }
+}
