+++ src/main/git/source-control/git-runtime-options-cache-key.ts
@@ -1,0 +1,5 @@
+import type { GitRuntimeOptions } from '../git-runtime-options'
+
+export function gitRuntimeOptionsKey(options: GitRuntimeOptions): readonly unknown[] {
+  return [options.wslDistro ?? null]
+}
