+++ src/main/git/source-control/git-show-max-bytes.ts
@@ -1,0 +1,2 @@
+export const MAX_GIT_SHOW_BYTES = 10 * 1024 * 1024
+export const MAX_STAGED_COMMIT_CONTEXT_BYTES = MAX_GIT_SHOW_BYTES
