+++ src/main/git/source-control/staging.ts
@@ -1,0 +1,99 @@
+import type { GitRuntimeOptions } from '../git-runtime-options'
+import { gitOptionsForWorktree } from '../git-runtime-options'
+import { gitExecFileAsync } from '../runner'
+import { invalidateGitReadCaches } from './git-read-cache-invalidation'
+import { BULK_CHUNK_SIZE, literalPathspec } from './git-pathspec'
+
+/**
+ * Stage a file.
+ */
+export async function stageFile(
+  worktreePath: string,
+  filePath: string,
+  options: GitRuntimeOptions = {}
+): Promise<void> {
+  invalidateGitReadCaches()
+  try {
+    await gitExecFileAsync(
+      ['add', '--', literalPathspec(filePath, options)],
+      gitOptionsForWorktree(worktreePath, options)
+    )
+  } finally {
+    invalidateGitReadCaches()
+  }
+}
+
+/**
+ * Unstage a file.
+ */
+export async function unstageFile(
+  worktreePath: string,
+  filePath: string,
+  options: GitRuntimeOptions = {}
+): Promise<void> {
+  invalidateGitReadCaches()
+  try {
+    await gitExecFileAsync(['restore', '--staged', '--', literalPathspec(filePath, options)], {
+      ...gitOptionsForWorktree(worktreePath, options)
+    })
+  } finally {
+    invalidateGitReadCaches()
+  }
+}
+
+/**
+ * Bulk stage files in batches to avoid E2BIG.
+ */
+export async function bulkStageFiles(
+  worktreePath: string,
+  filePaths: string[],
+  options: GitRuntimeOptions = {}
+): Promise<void> {
+  invalidateGitReadCaches()
+  if (filePaths.length === 0) {
+    return
+  }
+  try {
+    for (let i = 0; i < filePaths.length; i += BULK_CHUNK_SIZE) {
+      const chunk = filePaths.slice(i, i + BULK_CHUNK_SIZE)
+      await gitExecFileAsync(
+        ['add', '--', ...chunk.map((filePath) => literalPathspec(filePath, options))],
+        gitOptionsForWorktree(worktreePath, options)
+      )
+    }
+  } finally {
+    invalidateGitReadCaches()
+  }
+}
+
+/**
+ * Bulk unstage files in batches to avoid E2BIG.
+ */
+export async function bulkUnstageFiles(
+  worktreePath: string,
+  filePaths: string[],
+  options: GitRuntimeOptions = {}
+): Promise<void> {
+  invalidateGitReadCaches()
+  if (filePaths.length === 0) {
+    return
+  }
+  try {
+    for (let i = 0; i < filePaths.length; i += BULK_CHUNK_SIZE) {
+      const chunk = filePaths.slice(i, i + BULK_CHUNK_SIZE)
+      await gitExecFileAsync(
+        [
+          'restore',
+          '--staged',
+          '--',
+          ...chunk.map((filePath) => literalPathspec(filePath, options))
+        ],
+        {
+          ...gitOptionsForWorktree(worktreePath, options)
+        }
+      )
+    }
+  } finally {
+    invalidateGitReadCaches()
+  }
+}
