+++ src/main/git/source-control/submodule-diff.ts
@@ -1,0 +1,80 @@
+import type { GitDiffResult } from '../../../shared/git-diff-compare-types'
+import type { GitRuntimeOptions } from '../git-runtime-options'
+import { buildDiffResult } from './diff-result'
+import { readGitBlobAtOidPath } from './git-blob-read'
+import { resolveSubmoduleWorktreePath } from './submodule-paths'
+import {
+  readGitlinkOidFromIndex,
+  readGitlinkOidFromTree,
+  readWorkingSubmoduleHead
+} from './submodule-gitlink-oid'
+
+/**
+ * Synthesize a gitlink pointer diff: Git represents submodule commit changes as a
+ * one-line `Subproject commit <oid>` swap, so the old/new oids feed the text differ.
+ */
+export async function buildSubmodulePointerDiff(
+  worktreePath: string,
+  submodulePath: string,
+  staged: boolean,
+  compareAgainstHead: boolean,
+  options: GitRuntimeOptions,
+  // Why: default to the validated resolver so every caller is guarded against path escape.
+  submoduleWorktreePath = resolveSubmoduleWorktreePath(worktreePath, submodulePath)
+): Promise<GitDiffResult> {
+  let leftOid = ''
+  let rightOid = ''
+  if (staged) {
+    leftOid = await readGitlinkOidFromTree(worktreePath, 'HEAD', submodulePath, options)
+    rightOid = await readGitlinkOidFromIndex(worktreePath, submodulePath, options)
+  } else if (compareAgainstHead) {
+    leftOid = await readGitlinkOidFromTree(worktreePath, 'HEAD', submodulePath, options)
+    rightOid = await readWorkingSubmoduleHead(submoduleWorktreePath, options)
+  } else {
+    leftOid =
+      (await readGitlinkOidFromIndex(worktreePath, submodulePath, options)) ||
+      (await readGitlinkOidFromTree(worktreePath, 'HEAD', submodulePath, options))
+    rightOid = await readWorkingSubmoduleHead(submoduleWorktreePath, options)
+  }
+  return buildDiffResult(
+    leftOid ? `Subproject commit ${leftOid}\n` : '',
+    rightOid ? `Subproject commit ${rightOid}\n` : '',
+    false,
+    false,
+    submodulePath
+  )
+}
+
+/**
+ * Diff a file inside a submodule across two of its commits — used when the parent
+ * gitlink moved but the submodule worktree is clean (change is committed).
+ */
+export async function buildSubmoduleInnerCommitRangeDiff(
+  submoduleWorktreePath: string,
+  innerPath: string,
+  fromOid: string,
+  toOid: string,
+  options: GitRuntimeOptions
+): Promise<GitDiffResult> {
+  let originalContent = ''
+  let modifiedContent = ''
+  let originalIsBinary = false
+  let modifiedIsBinary = false
+  try {
+    const left = await readGitBlobAtOidPath(submoduleWorktreePath, fromOid, innerPath, options)
+    originalContent = left.content
+    originalIsBinary = left.isBinary
+    const right = await readGitBlobAtOidPath(submoduleWorktreePath, toOid, innerPath, options)
+    modifiedContent = right.content
+    modifiedIsBinary = right.isBinary
+  } catch {
+    // Fallback to empty content; a missing blob (add/delete) reads as one side.
+  }
+  return buildDiffResult(
+    originalContent,
+    modifiedContent,
+    originalIsBinary,
+    modifiedIsBinary,
+    innerPath
+  )
+}
